Move magnification types to winBindings
